XIVY-3568 split primeui and ivyWebTest annotation
- change package names
- change versions (primeui to 7.0.0-SNAPSHOT, webtest to 8.0.0-SNAPSHOT)
