+++ .gitignore
@@ -1,0 +1,15 @@
+# Eclipse
+.settings/
+
+# Java
+.class
+
+# Mac
+.DS_Store
+
+# Maven
+/log/
+target/
+
+# VS Code
+.vscode/
