--- primeui-tester/.gitignore
@@ -1,12 +1,0 @@
-# Eclipse
-.settings/
-
-# Java
-.class
-
-# Mac
-.DS_Store
-
-# Maven
-/log/
-target/
